backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
torch
matplotlib
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
